rename folder to network
