+++ pyvipr/network_viz/__init__.py
@@ -1,0 +1,3 @@
+from pyvipr.network_viz.views import *
+
+__all__ = views.__all__
